requirements.txt
requests
sanic
python-twitter
bs4
boto3